Update 2025-03-13-trevolution.md
There is a bit which does  not make sense in the first sentence. I have removed the description/definition of a TRE " are highly secure,and work on approved projects of public benefit.)" because it does not fit there, grammatically. Does it need to be included somewhere else still? I see the UoM news item took this sentence out too.